requirements.txt
opencv-python==4.8.1
pillow==10.0.0
tkcalendar==1.6.1 